requirements.txt
botocore==1.18.5
Brotli==1.0.9
cache==1.0.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
future==0.18.2
idna==2.10
itsdangerous==1.1.0
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numba==0.51.2
numpy==1.19.2
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.9.27
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
six==1.15.0
threadpoolctl==2.1.0
torch
tqdm==4.49.0
umap-learn==0.4.6
urllib3==1.25.10
moviepy==1.0.3